server/requirements.txt
certifi
fastapi
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchain-groq
langgraph
pydantic
pydantic_core
python-dotenv
pymongo
SQLAlchemy
uvicorn
python-telegram-bot
httpx

